Namespace spec helpers for DfE Sign-in
