+++ spec/rails_helper.rb
@@ -64,6 +64,7 @@ RSpec.configure do |config|
   config.include FeatureHelpers, type: :feature
   config.include ActiveSupport::Testing::TimeHelpers
   config.include ActiveJob::TestHelper
+  config.include DfeSignInHelpers
 
   config.before :each do
     clear_enqueued_jobs
